#11 :hammer: 사진 목록 보여주는 뷰 구현함
